refactor: remove unused imports, modulize text
